Consolidate duplicated exam observation, template, and modal helpers into shared utilities
- Extract extractObservationsFromExamComponents and collectKnownExamFields to packages/utils for shared use by review page, PDF, and admin templates
- Add L/R paired modal-exam handling to PDF to match review page display
- Extract buildAllOptions, buildDescriptionMap, buildAbnormalMap to exam-modal-helpers for ExamModalCheckbox and ExamPairedModalCheckbox
- Move template detail types (TemplateExamFinding, TemplateCodeInfo, etc.) to packages/utils/lib/types/data/list-template.types.ts
- Extract verifyIsTemplate and findHolderList to shared template-helpers for admin zambdas
