+++ apps/ehr/src/features/visits/telemed/components/admin/GlobalTemplateDetailPage.tsx
@@ -27,48 +27,7 @@ import CustomBreadcrumbs from 'src/components/CustomBreadcrumbs';
 import { QUERY_STALE_TIME } from 'src/constants';
 import { useApiClients } from 'src/hooks/useAppClients';
 import PageContainer from 'src/layout/PageContainer';
-
-interface ExamFinding {
-  fieldName: string;
-  label: string;
-  isAbnormal: boolean;
-  note: string;
-}
-
-interface DiagnosisInfo {
-  code: string;
-  display: string;
-}
-
-interface CodeInfo {
-  code: string;
-  display: string;
-}
-
-interface TemplateDetailData {
-  templateName: string;
-  templateId: string;
-  examVersion: string;
-  isCurrentVersion: boolean;
-  sections: {
-    hpiNote: string | null;
-    moiNote: string | null;
-    rosNote: string | null;
-    examFindings: ExamFinding[];
-    mdm: string | null;
-    diagnoses: DiagnosisInfo[];
-    patientInstructions: string | null;
-    cptCodes: CodeInfo[];
-    emCode: CodeInfo | null;
-    accident: {
-      autoAccident: boolean;
-      employment: boolean;
-      otherAccident: boolean;
-      date?: string;
-      state?: string;
-    } | null;
-  };
-}
+import { TemplateDetailOutput } from 'utils';
 
 function renderMarkdown(text: string): ReactElement {
   // Convert markdown task lists and basic formatting to HTML-like rendering
@@ -152,13 +111,13 @@ export default function GlobalTemplateDetailPage(): ReactElement {
   const navigate = useNavigate();
   const { oystehrZambda } = useApiClients();
 
-  const { data, isLoading, error } = useQuery<TemplateDetailData, Error>({
+  const { data, isLoading, error } = useQuery<TemplateDetailOutput, Error>({
     queryKey: ['template-detail', templateId],
     queryFn: async () => {
       if (!oystehrZambda || !templateId) {
         throw new Error('API client or template ID not available');
       }
-      return (await getTemplateDetail(oystehrZambda, { templateId })) as unknown as TemplateDetailData;
+      return (await getTemplateDetail(oystehrZambda, { templateId })) as unknown as TemplateDetailOutput;
     },
     enabled: !!oystehrZambda && !!templateId,
     staleTime: QUERY_STALE_TIME,
