+++ packages/zambdas/src/ehr/list-templates/index.ts
@@ -6,7 +6,6 @@ import {
   ExamType,
   getSecret,
   GLOBAL_TEMPLATE_IN_PERSON_CODE_SYSTEM,
-  GLOBAL_TEMPLATE_META_TAG_CODE_SYSTEM,
   GLOBAL_TEMPLATE_TELEMED_CODE_SYSTEM,
   ListTemplatesZambdaInput,
   ListTemplatesZambdaOutput,
@@ -15,6 +14,7 @@ import {
 } from 'utils';
 import { checkOrCreateM2MClientToken, topLevelCatch, wrapHandler, ZambdaInput } from '../../shared';
 import { createOystehrClient } from '../../shared/helpers';
+import { findHolderList } from '../shared/template-helpers';
 import { validateRequestParameters } from './validateRequestParameters';
 
 // Lifting up value to outside of the handler allows it to stay in memory across warm lambda invocations
@@ -46,17 +46,8 @@ const performEffect = async (
 ): Promise<ListTemplatesZambdaOutput> => {
   const { examType } = validatedInput;
 
-  // Find the holder list (tagged with GLOBAL_TEMPLATE_META_TAG_CODE_SYSTEM)
-  const holderLists = (
-    await oystehr.fhir.search<List>({
-      resourceType: 'List',
-      params: [{ name: '_tag', value: `${GLOBAL_TEMPLATE_META_TAG_CODE_SYSTEM}|` }],
-    })
-  ).unbundle();
-
-  const holderList = holderLists.find(
-    (l) => l.meta?.tag?.some((tag) => tag.system === GLOBAL_TEMPLATE_META_TAG_CODE_SYSTEM)
-  );
+  // Find the holder list
+  const holderList = await findHolderList(oystehr);
 
   if (!holderList || !holderList.entry?.length) {
     return { templates: [] };
